Fixed an errors when calling class
FIXED: in the initial Monster class, I changed self.level = 1 to self.level = level. My suspicion is that this is meant to be as abstracted as possible, and giving real values (i.e. 1, 2, 100) actually made the variable concrete and non-malleable. Setting it to 'level' allows the iteration to call the class and modify it, whereas having a 'real' value does not allow for modification.